move all canvas editing code into editing, and split into multiple files
